Non far from production

+++ src/View/SwingModules/List.java
@@ -11,7 +11,7 @@ import java.util.ArrayList;
 /**
  * List data in table
  */
-public class List extends JPanel {
+public class List extends JPanel implements BuilderInterface {
     ArrayList<Entity> entities;
     // Table for user data
     private final JTable table;
@@ -20,7 +20,7 @@ public class List extends JPanel {
 
     // buttons
     public BackButton backButton = new BackButton();
-    JButton delete = new JButton("Supprimer la sélection");
+    JButton delete = new JButton("Supprimer");
 
     public List(String[] tableColumn) {
         add(backButton.getToolBar());
@@ -33,11 +33,16 @@ public class List extends JPanel {
         JScrollPane userTableScroll = new JScrollPane(table, JScrollPane.VERTICAL_SCROLLBAR_ALWAYS,
             JScrollPane.HORIZONTAL_SCROLLBAR_ALWAYS);
         add(userTableScroll);
+    }
 
-        delete.setSize(400, 50);
-        add(delete, BoxLayout.LINE_AXIS);
+    @Override
+    public List create(JPanel jpanel) {
+        if (delete != null) {
+            addButton(delete);
+            this.delete(e -> this.handleDelete());
+        }
 
-        this.delete(e -> this.handleDelete());
+        return this;
     }
 
     // gets data from database and loads to table
@@ -55,6 +60,7 @@ public class List extends JPanel {
     }
 
     protected void handleDelete() {
+        if (!this.confirm()) return;
         if (table.getSelectedRow() > -1) {
             // remove selected row from the model
             DefaultTableModel model = (DefaultTableModel) table.getModel();
@@ -62,8 +68,14 @@ public class List extends JPanel {
             int selectedRow = table.getSelectedRow();
             model.removeRow(selectedRow);
             this.entities.remove(selectedRow);
-            JOptionPane.showMessageDialog(null, "Suppresion effectuée");
         }
+    }
+
+    public Entity getEntity() {
+        if (table.getSelectedRow() <= -1) return null;
+        int selectedRow = table.getSelectedRow();
+
+        return this.entities.get(selectedRow);
     }
 
     public void update(TableModelListener listener) {
@@ -74,7 +86,35 @@ public class List extends JPanel {
         delete.addActionListener(actionListener);
     }
 
+    public List disableDeleteButton()
+    {
+        delete = null;
+
+        return this;
+    }
+
+    public List addButton(JButton button) {
+        button.setSize(400, 50);
+        add(button, BoxLayout.LINE_AXIS);
+
+        return this;
+    }
+
     public JTable getTable() {
         return table;
+    }
+
+    /**
+     * Show confirm dialog.
+     *
+     * @return boolean true user answer Yes.
+     */
+    public boolean confirm() {
+        return JOptionPane.showConfirmDialog(
+            this,
+            "Supprimer ?",
+            "Confirmer la suppression",
+            JOptionPane.YES_NO_OPTION
+        ) == 0;
     }
 }
